requirements.txt
biopython
requests
